Implementation du graph selon la popularite
